requirements.txt
redis>=5.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
streamlit>=1.30.0
python-telegram-bot>=21.0
langfuse
# Gmail API
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
qdrant-client>=1.7.0
fastapi>=0.109.0
uvicorn>=0.27.0
